+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Nav from "./atoms/Nav";
+import Nav from "./molecules/Nav";
 import Logo from "./atoms/Logo";
 import LoginButton from "./atoms/LoginButton";
 
@@ -9,10 +9,8 @@ function Header() {
       <Logo />
       {/* 로그인버튼을 atom으로 빼서 Nav안에 넣는게 좀 이상하지만 스타일 적용에 어려움이 생겨 넣었음.. */}
       <Nav>
-        <Nav.List>
-          <Nav.Item to="upload" body="내 물건 팔기" />
-          <Nav.Item to="profile/:id" body="내 정보" />
-        </Nav.List>
+        <Nav.Item to="upload" body="내 물건 팔기" />
+        <Nav.Item to="profile/:id" body="내 정보" />
         <LoginButton />
       </Nav>
     </HeaderLayout>
